requirements.txt
python-dotenv
jupyter
anthropic
langgraph
langchain-core